Refactor & Docs: 치명적 레거시 버그 수정, 보안 마이그레이션 및 README 작성
[Security] 비밀번호 평문 저장(BCrypt) 및 식별자 충돌 이슈(UUID) 해결
[Security] 하드코딩된 민감정보(DB 계정, OAuth2 Client Secret) 환경변수로 격리
[Fix] 주관식 설문조사 시 덮어쓰기로 인한 데이터 유실 버그(CsService) 해결
[Fix] 과거 병합 과정에서 남은 Git Conflict 마커 제거 및 코드 정합성 복구
[Refactor] 자바 네이밍 컨벤션 위배 및 오타 컨트롤러 리팩토링 (AdminController, LoginController)
[Feat] API 명세화 및 테스트 구조 개선을 위한 Swagger(springdoc-openapi) 의존성 및 설정 추가
[Docs] 파편화된 서비스 아키텍처에 대한 원인(팀 병합 잔재) 분석 및 향후 적용할 도메인 주도 설계(SRP, Facade) 회고록 README 작성

+++ counsel/src/main/java/com/example/counsel/repository/UserRepository.java
@@ -3,11 +3,9 @@ package com.example.counsel.repository;
 import com.example.counsel.entity.UserEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-
-public interface UserRepository extends JpaRepository<UserEntity, Integer> {
+public interface UserRepository extends JpaRepository<UserEntity, Long> {
 
     boolean existsByUsername(String username);
-
 
     UserEntity findByUsername(String username);
